+++ client/acp/components/Donate.react.js
@@ -1,68 +1,25 @@
-var React            = require('react'),
-    LinkedStateMixin = require('react/lib/LinkedStateMixin'),
-    App              = require('app'),
-    StripeCheckout   = require('StripeCheckout');
+var React = require('react');
 
 var Donate = React.createClass({
-    mixins: [LinkedStateMixin],
-
-    componentDidMount: function () {
-        this.stripeHandler = StripeCheckout.configure({
-            key       : 'pk_live_AcfQs725nv7nIF5sRCG3v4Q8',
-            image     : 'https://s3.amazonaws.com/stripe-uploads/acct_16mDSJB8UmE70jk7merchant-icon-1442539384457-ava-mdpi.jpg',
-            locale    : 'auto',
-            panelLabel: 'Donate {{amount}}',
-            email     : App.user.email,
-            bitcoin   : true,
-            token     : function (token) {
-                // Use the token to create the charge with a server-side script.
-                // You can access the token ID with `token.id`
-                // NOOP
-            }
-        });
-    },
-
-    getInitialState: function () {
-        return {
-            amount: null
-        };
-    },
 
     render: function () {
         return (
             <div className="panel panel-default">
                 <div className="panel-body">
-                    <p>Do you like plugin? Support developer. Make a donation. Thank you in advance.</p>
+                    <p>Do you like a plugin? Make a donation, It will definitely help as an extra motivation to push new
+                        features. Thank you.</p>
 
-                    <div className="input-group">
-                        <input
-                            type="text"
-                            className="form-control"
-                            placeholder="Amount"
-                            valueLink={this.linkState('amount')}/>
-                        <span className="input-group-btn">
-                            <button
-                                className="btn btn-primary"
-                                type="button"
-                                onClick={this._donateDidClick}>Donate <small>via Stripe</small>
-                            </button>
-                        </span>
-                    </div>
+                    <form action="https://www.paypal.com/cgi-bin/webscr" method="post" target="_top">
+                        <input type="hidden" name="cmd" value="_s-xclick"/>
+                        <input type="hidden" name="encrypted" value="-----BEGIN PKCS7-----MIIHRwYJKoZIhvcNAQcEoIIHODCCBzQCAQExggEwMIIBLAIBADCBlDCBjjELMAkGA1UEBhMCVVMxCzAJBgNVBAgTAkNBMRYwFAYDVQQHEw1Nb3VudGFpbiBWaWV3MRQwEgYDVQQKEwtQYXlQYWwgSW5jLjETMBEGA1UECxQKbGl2ZV9jZXJ0czERMA8GA1UEAxQIbGl2ZV9hcGkxHDAaBgkqhkiG9w0BCQEWDXJlQHBheXBhbC5jb20CAQAwDQYJKoZIhvcNAQEBBQAEgYAlPB9ODJp/8PkSqyISzAfSnq4BSEh8eZeDZ2l2IZU8q1vZGt/+2oRvosyih6gjCvbPV9am3taCAg9Izqam40SuxP56c0QpIyZC0PxARsivhFCJiBR0SqYtHxUFKjY6jK95NQIkIa/b97cNvcT1QJWVD0nKbaV9rwxtP1xw7W0UHjELMAkGBSsOAwIaBQAwgcQGCSqGSIb3DQEHATAUBggqhkiG9w0DBwQI2dngr3vXGruAgaB9HZJetUwuQEtdHzIs8UDrlxYIBPL/9dv8/1uOZ7Qq6sKUuXORhIk68lM4glytEkJsVc+dGG6sAe+QovC3sb/qBby4lgI8XQlx2DQwE8XPYEv9tPtWuCr6XVWbWQb30PcPhdqci3e4bIvie6rne+Snxa8m9OpnFlarNrkyot+YSkH1ZL9rKg7YHhzGxMIzsP67ZACNxdgbRrOKG7ETgENooIIDhzCCA4MwggLsoAMCAQICAQAwDQYJKoZIhvcNAQEFBQAwgY4xCzAJBgNVBAYTAlVTMQswCQYDVQQIEwJDQTEWMBQGA1UEBxMNTW91bnRhaW4gVmlldzEUMBIGA1UEChMLUGF5UGFsIEluYy4xEzARBgNVBAsUCmxpdmVfY2VydHMxETAPBgNVBAMUCGxpdmVfYXBpMRwwGgYJKoZIhvcNAQkBFg1yZUBwYXlwYWwuY29tMB4XDTA0MDIxMzEwMTMxNVoXDTM1MDIxMzEwMTMxNVowgY4xCzAJBgNVBAYTAlVTMQswCQYDVQQIEwJDQTEWMBQGA1UEBxMNTW91bnRhaW4gVmlldzEUMBIGA1UEChMLUGF5UGFsIEluYy4xEzARBgNVBAsUCmxpdmVfY2VydHMxETAPBgNVBAMUCGxpdmVfYXBpMRwwGgYJKoZIhvcNAQkBFg1yZUBwYXlwYWwuY29tMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDBR07d/ETMS1ycjtkpkvjXZe9k+6CieLuLsPumsJ7QC1odNz3sJiCbs2wC0nLE0uLGaEtXynIgRqIddYCHx88pb5HTXv4SZeuv0Rqq4+axW9PLAAATU8w04qqjaSXgbGLP3NmohqM6bV9kZZwZLR/klDaQGo1u9uDb9lr4Yn+rBQIDAQABo4HuMIHrMB0GA1UdDgQWBBSWn3y7xm8XvVk/UtcKG+wQ1mSUazCBuwYDVR0jBIGzMIGwgBSWn3y7xm8XvVk/UtcKG+wQ1mSUa6GBlKSBkTCBjjELMAkGA1UEBhMCVVMxCzAJBgNVBAgTAkNBMRYwFAYDVQQHEw1Nb3VudGFpbiBWaWV3MRQwEgYDVQQKEwtQYXlQYWwgSW5jLjETMBEGA1UECxQKbGl2ZV9jZXJ0czERMA8GA1UEAxQIbGl2ZV9hcGkxHDAaBgkqhkiG9w0BCQEWDXJlQHBheXBhbC5jb22CAQAwDAYDVR0TBAUwAwEB/zANBgkqhkiG9w0BAQUFAAOBgQCBXzpWmoBa5e9fo6ujionW1hUhPkOBakTr3YCDjbYfvJEiv/2P+IobhOGJr85+XHhN0v4gUkEDI8r2/rNk1m0GA8HKddvTjyGw/XqXa+LSTlDYkqI8OwR8GEYj4efEtcRpRYBxV8KxAW93YDWzFGvruKnnLbDAF6VR5w/cCMn5hzGCAZowggGWAgEBMIGUMIGOMQswCQYDVQQGEwJVUzELMAkGA1UECBMCQ0ExFjAUBgNVBAcTDU1vdW50YWluIFZpZXcxFDASBgNVBAoTC1BheVBhbCBJbmMuMRMwEQYDVQQLFApsaXZlX2NlcnRzMREwDwYDVQQDFAhsaXZlX2FwaTEcMBoGCSqGSIb3DQEJARYNcmVAcGF5cGFsLmNvbQIBADAJBgUrDgMCGgUAoF0wGAYJKoZIhvcNAQkDMQsGCSqGSIb3DQEHATAcBgkqhkiG9w0BCQUxDxcNMTcwNjIwMDI1OTUyWjAjBgkqhkiG9w0BCQQxFgQUQN4AlBjPPtsSO6/AdtP6b9K4CVYwDQYJKoZIhvcNAQEBBQAEgYBU3se3kgahCBBM9T6P8HSwbVOf+t60Y1r913qWzJG+QCJnSWfM2IByb4DGU8MPkUHUAqKvPa+JQevYTMRkght7BKDdtNgvO7ddy82kTY0ImfSBdZoeVQ3c+cQeC78XmMh2zaAHcJVtgUtHBFMuuqHYUsuJxrEtYlLauZrVlTXTDg==-----END PKCS7-----"/>
+                        <input type="image" src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif" border="0" name="submit" alt="PayPal - The safer, easier way to pay online!"/>
+                        <img alt="" border="0" src="https://www.paypalobjects.com/en_US/i/scr/pixel.gif" width="1" height="1"/>
+                    </form>
                 </div>
             </div>
         );
-    },
-
-    _donateDidClick: function () {
-        var amount = parseFloat(this.state.amount) * 100;
-        amount = amount || 500;
-
-        this.stripeHandler.open({
-            name       : 'Nicolas Siver',
-            description: 'NS Custom Fields Donation',
-            amount     : amount
-        });
     }
+
 });
 
 module.exports = Donate;