+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v5.0.0 - 06/19/2017
+
+- Compatibility with NodeBB v1.5.0
+- Remove extra payment method from a Donation section
+- Update development dependencies for ACP panel
+
 ## v4.0.0 - 03/09/2016
 
 - Use ACP Scripts capabilities of NodeBB